Use "LearnToCode" category
This non-standard category is used to populate the "Learn to Code" tab
in the Endless OS app center.

While it's true that LearnToCode is not a standard category name, and so
(by the spec) we should be using the X- prefix for a vendor extension,
in practice X-prefixed categories do not appear in the Flathub
appstream, and the app center only looks for LearnToCode anyway.